% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_overrides.R
\name{require}
\alias{require}
\alias{q}
\alias{quit}
\alias{Sys.setlocale}
\alias{setwd}
\title{Overrides for base functions}
\usage{
require(package, quietly = FALSE, character.only = FALSE, ...)

q(save = "default", status = 0, runLast = TRUE, ...)

quit(save = "default", status = 0, runLast = TRUE, ...)

Sys.setlocale(category = "LC_ALL", locale = "", ...)

setwd(dir)
}
\arguments{
\item{package}{see \code{\link[base:require]{require}}}

\item{quietly}{see \code{\link[base:require]{require}}}

\item{character.only}{see \code{\link[base:require]{require}}}

\item{...}{additional arguments, see respective base function.}

\item{save}{see \code{\link[base:q]{q}}}

\item{status}{see \code{\link[base:q]{q}}}

\item{runLast}{see \code{\link[base:q]{q}}}

\item{category}{see \code{\link[base:Sys.setlocale]{Sys.setlocale}}}

\item{locale}{see \code{\link[base:Sys.setlocale]{Sys.setlocale}}}

\item{dir}{see \code{\link[base:setwd]{setwd}}}
}
\value{
See \code{\link[base:require]{require}}, \code{\link[base:q]{q}},
  \code{\link[base:Sys.setlocale]{Sys.setlocale}}, or \code{\link[base:setwd]{setwd}}.
}
\description{
RKWard replaces these functions seamlessly with wrappers that add specific
features, i.e. to integrate well with the GUI:
}
\details{
\itemize{
  \item{\code{\link[base:require]{require}}}
  \item{\code{\link[base:q]{q/quit}}}
  \item{\code{\link[base:Sys.setlocale]{Sys.setlocale}}}
  \item{\code{\link[base:setwd]{setwd}}}
}
}
