% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.utility-functions.R
\name{rk.adjust.system.path}
\alias{rk.adjust.system.path}
\title{Add one or more paths to the filesystem search path used in this session}
\usage{
rk.adjust.system.path(add)
}
\arguments{
\item{add}{Paths to add. May be missing, in which case the path will no be touched.}
}
\value{
A vector of the directories in the file sytem path after the adjustment
}
\description{
Add the given path to the "PATH" environment variable of the running R session. This
can be useful to make sure external binaries are found by Sys.which. Paths are normalized
before being added, and duplicates are stripped from the path.
}
