% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.check_for_pandoc.R
\name{rk.check.for.pandoc}
\alias{rk.check.for.pandoc}
\title{Check for pandoc installation and its features}
\usage{
rk.check.for.pandoc()
}
\value{
A named list with at least the boolean element \code{available}, and if
   that is \code{TRUE}, also the character vectors \code{path}, \code{output_formats},
      and \code{version}.
}
\description{
This function looks for a pandoc installation and if found, gathers information
on its path, supported output formats, and the version number. It has no parameters.
}
\examples{

rk.check.for.pandoc()
}
\seealso{
\code{\link{rk.sessionInfo}}
}
\author{
Meik Michalke \email{rkward-devel@kde.org}
}
\keyword{misc}
\keyword{utilities}
