% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.completion-functions.R
\name{.rk.completions}
\alias{.rk.completions}
\title{Helper functions for querying completions}
\usage{
.rk.completions(fragment, type)
}
\arguments{
\item{fragment}{string fragment to complete.}

\item{type}{one of \code{"funargs"}, \code{"$"}, \code{"@"}, \code{"::"}, \code{":::"}, or \code{"?"}.}
}
\description{
These are currently not exported, as they are not intended to be use in user code
The API may or may not be stable.
}
