% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_debugger.R
\name{rk.relative.src.line}
\alias{rk.relative.src.line}
\title{Get relative source location}
\usage{
rk.relative.src.line(inner, outer)
}
\arguments{
\item{inner}{either an object containing source references, or an object of class
\code{\link[base:srcref]{srcref}}.}

\item{outer}{see \code{inner}.}
}
\value{
An integer or \code{NA}.
}
\description{
Get relative source location
}
\seealso{
\code{\link[utils:getSrcref]{getSrcref}},
  \code{\link[utils:getSrcFilename]{getSrcFilename}},
  \code{\link[utils:getSrcLocation]{getSrcLocation}}
}
