% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.preview-functions.R
\name{rk.render.markdown.preview}
\alias{rk.render.markdown.preview}
\title{Render a preview of the given R markdown file}
\usage{
rk.render.markdown.preview(infile, outdir, ...)
}
\arguments{
\item{infile}{The input Rmd file, specified as a character string.}

\item{outdir}{The directory for output. The directory is not cleared, but existing files inside it may be overwritten without notice!}

\item{...}{Additional parameters to pass to \link{rmarkdown::render}. Notably this may include \code{utput_format="html_document"} or
\code{output_format="pdf_document"}.}
}
\description{
This function is mostly targetted at providing a preview inside a script editor window. For other purposes
it will generally make more sense to call \link{rmarkdown::render}, directly.
}
