% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.utility-functions.R
\name{rk.switch.frontend.language}
\alias{rk.switch.frontend.language}
\title{Switch language / translation to use in the frontend.}
\usage{
rk.switch.frontend.language(LANG = "C")
}
\arguments{
\item{LANG}{language code to use. "C" for no translation, i.e. generally English}
}
\description{
This feature is mostly intended for the purpose of automated testing, which needs a
defined language to work. It might also be useful for translators, or e.g. to look up
some terms untranslated in special cases where you are more familiar with the English terms than
your native language terms. Note that this will only strings that are translated after the call., only those which get
translated after the call. Most new dialogs you open, and importantly new plugin dialogs should
show strings in the new lanuage, however.
}
\details{
To change the language in the backend, use \code{Sys.setenv(LANGUAGE=...)} or \code{Sys.setlocale()}.
}
