\name{rk.sync}
\alias{rk.sync}
\alias{rk.sync.global}

\title{TODO}

\usage{
rk.sync(x)
rk.sync.global()
}

\arguments{
}

\details{
}

\value{
}

\author{Thomas Friedrichsmeier \email{rkward-devel@lists.sourceforge.net}}

\section{Warning}{
This is not to be used by a normal user.
}


\examples{
}

\keyword{utilities}
\keyword{misc}
