% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.workspace-functions.R
\name{.rk.with.placement.hint}
\alias{.rk.with.placement.hint}
\alias{rk.with.placement.hint}
\title{Control window placement}
\usage{
.rk.with.placement.hint(hint, expr)
}
\arguments{
\item{hint}{a character string specifying the placement in the form "[attached|detached][:name]"
If a name is given, and this position is not yet known, the attached/detached hint will
be followed (or, if ommitted, default placement will be used). If later a second window is
created with the same given name, it will replace the first window.}

\item{expr}{Expression to evaluate, unsually an expression that is expected to create exactly one
new window.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{.rk.with.placement.hint} can be used to make windows appear in a specific
location: attached, detached, or in a named position where a previous window is
found. (The latter used for preview windows, importantly).
}
\details{
NOTE: This function is still somewhat experimental, and it is not guaranteed that
it will remain in place, with compatible parameters.
}
\examples{

## Not run
.rk.with.placement.hint ("attached", {
   RK ()
   plot (1, 1)
})
## End not run

}
\author{
Thomas Friedrichsmeier \email{rkward-devel@kde.org}
}
\keyword{utilities}

