% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_graphics.R
\name{rk.without.plot.history}
\alias{rk.without.plot.history}
\title{Run a (plotting) action, without recording anything in the plot history.
Internally, the plot history option is turned off for the duration of the action.}
\usage{
rk.without.plot.history(expr)
}
\description{
Run a (plotting) action, without recording anything in the plot history.
Internally, the plot history option is turned off for the duration of the action.
}

