\name{rk.call.plugin}
\alias{rk.call.plugin}

\title{TODO}

\description{
}

\usage{
rk.call.plugin(plugin, ..., submit.mode = c("manual", "auto", "submit"))
}

\arguments{
}

\details{
}

\value{
  Returns \code{TRUE} invisibly.
}

\author{Thomas Friedrichsmeier \email{rkward-devel@lists.sourceforge.net}}

\examples{
}

\keyword{utilities}
