\name{RKWard_misc}
\alias{rk.misc}
\alias{rk.rename.in.container}
\alias{rk.make.repos.string}
\alias{rk.select.CRAN.mirror}

\title{TODO}

\description{
}

\usage{
rk.rename.in.container(x, old_name, new_name, envir = parent.frame())
rk.make.repos.string()
rk.select.CRAN.mirror()
}

\arguments{
}

\details{
}

\value{
}

\author{Thomas Friedrichsmeier \email{rkward-devel@lists.sourceforge.net}}

\examples{
}

\keyword{utilities}
