% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rk.old.packages}
\alias{rk.old.packages}
\title{Slightly smarter variant of old.packages()}
\usage{
rk.old.packages(lib.loc = NULL, repos = getOption("repos"),
  contriburl = contrib.url(repos, type),
  instPkgs = installed.packages(lib.loc = lib.loc), method,
  available = NULL, checkBuilt = FALSE, type = getOption("pkgType"))
}
\value{
a character vector of packages which are really old
}
\description{
For most purposes, this function is identical to old.packages(). However, if the same
package is installed to different locations, in different versions, old.packages() will
treat each of these installations separately. Thus, e.g. if lib.loc == c("A", "B") and
package X is installed in B at an outdated version 0.1, but in A at the most recent version 0.2,
old.packages() will report package X at B as old. In contrast rk.old.packages() will recognize
that the current version is higher up in the path, and not report package X as old.
}
\examples{
## NOT RUN
rk.old.packages()
}
\author{
Thomas Friedrichsmeier \email{rkward-devel@lists.sourceforge.net}
}
\keyword{attribute}
\keyword{misc}
\keyword{utilities}

