\name{rk.results}
\alias{rk.print}
\alias{rk.print.literal}
\alias{rk.header}
\alias{rk.results}
\alias{rk.describe.alternative}

\title{TODO}

\description{
}

\usage{
rk.print(x, ...)
rk.print.literal(x)
rk.header(title, parameters = list(), level = 1)
rk.results(x, titles = NULL)
rk.describe.alternative(x)
}

\arguments{
}

\details{
}

\value{
}

\author{Thomas Friedrichsmeier \email{rkward-devel@lists.sourceforge.net}}

\seealso{
\url{rkward://page/rkward_output}
}

\examples{
}

\keyword{utilities}
