\name{rk.sync}
\alias{rk.sync}
\alias{rk.sync.global}

\title{Sync R object(s)}

\usage{
rk.sync(x)
rk.sync.global()
}

\arguments{
\item{x}{any R object to sync}
}

\details{
\bold{TODO}: Explain what is "sync"

While \code{rk.sync} syncs a single object, \code{x}, \code{rk.sync.global} syncs all objects in  \code{.GlobalEnv}.
}

\value{
\code{NULL}, invisibly.
}

\author{Thomas Friedrichsmeier \email{rkward-devel@lists.sourceforge.net}}

\section{Warning}{
A typical user should not need to use these.
}

\examples{
}

\keyword{utilities}
\keyword{misc}
