% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkward-package.R
\docType{package}
\name{rkward-package}
\alias{rkward}
\alias{rkward-package}
\title{\packageTitle{rkward}}
\description{
\packageDescription{rkward}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab rkward\cr
Type: \tab Package\cr
Version: \tab 0.8.0\cr
Date: \tab 2024-07-18\cr
Depends: \tab R (>= 2.9.0),methods\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab https://rkward.kde.org\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rkward.kde.org}
  \item Report bugs at \url{https://rkward.kde.org/Bugs.html}
}

}
\author{
\packageAuthor{rkward}

Maintainer: \packageMaintainer{rkward}
}
