% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rkward-package}
\alias{rkward}
\alias{rkward-package}
\title{Provides functions related to the RKWard GUI.}
\description{
Provides functions related to the RKWard GUI.
}
\details{
This package contains functions which are useful in combination with the RKWard GUI. Many of these
functions only needed for the internal communication between RKWard and R, but some are also useful in user scripts.
}
\author{
Thomas Friedrichsmeier \email{thomas.friedrichsmeier@ruhr-uni-bochum.de}, and the RKWard Team \email{rkward-devel@lists.sourceforge.net}
}
\keyword{package}

