% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rkward-package}
\alias{rkward-package}
\title{The rkward Package}
\description{
Provides functions related to the RKWard GUI.
}
\details{
\tabular{ll}{
Package: \tab rkward\cr
Type: \tab Package\cr
Version: \tab 0.6.3\cr
Date: \tab 2014-10-30\cr
Depends: \tab R (>= 2.9.0),methods\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab http://rkward.sourceforge.net\cr
}

This package contains functions which are useful in combination with the RKWard GUI. Many of these
functions only needed for the internal communication between RKWard and R, but some are also useful in user scripts.
}
\author{
Thomas Friedrichsmeier, with contributions from RKWard-devel mailing list
}
\keyword{package}

